+++ spec/ductwork/branch/advance_expand_spec.rb
@@ -133,14 +133,5 @@ RSpec.describe Ductwork::Branch, "#advance!" do
       expect(run.reload).to be_halted
       expect(run.pipeline).to be_halted
     end
-
-    it "halts all other branches" do
-      create(:branch, :in_progress, run:)
-
-      expect do
-        branch.advance!(transition, advancement)
-      end.not_to change(Ductwork::Step, :count)
-      expect(run.branches).to all(be_halted)
-    end
   end
 end